+++ tests/qs_tests.cpp
@@ -20,22 +20,22 @@
 #define CATCH_CONFIG_MAIN  // This tells Catch to provide a main() - only do this in one cpp file
 #include "catch2/catch.hpp"
 
-std::vector<int> getRandomList( int n )
+std::vector<int64_t> getRandomList( int64_t n )
 {
     std::random_device rd; // obtain a random number from hardware
     std::mt19937 gen(rd()); // seed the generator
     std::uniform_int_distribution<> distr(0, n*n); // define the range
 
-    std::vector<int> output;
+    std::vector<int64_t> output;
     output.reserve( n );
 
-    for(int i=0; i<n; ++i)
+    for(int64_t i=0; i<n; ++i)
         output.push_back( distr(gen) );
 
     return output;
 }
 
-void printList( const std::vector< int > & pivots ) {
+void printList( const std::vector< int64_t > & pivots ) {
     for( auto p : pivots )
         std::cout << p << ", ";
     std::cout << std::endl;
@@ -46,12 +46,12 @@ TEST_CASE( "qs::range", "" )
     using namespace std;
 
     SECTION( "inside_range" ) {
-        for( int np = 1; np <= 10000; ++np ) {
+        for( int64_t np = 1; np <= 10000; ++np ) {
             double total_duration = 0.0;
-            int num_durations = 0;
+            int64_t num_durations = 0;
             
-            for( int i = 1; i < 30; ++i ) {
-                std::vector< int > v = getRandomList( 300000 );
+            for( int64_t i = 1; i < 30; ++i ) {
+                std::vector< int64_t > v = getRandomList( 300000 );
 
                 std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now();
 
@@ -60,9 +60,9 @@ TEST_CASE( "qs::range", "" )
                 // std::qsort(
                 //     v.data(),
                 //     v.size(),
-                //     sizeof(int),
+                //     sizeof(int64_t),
                 //     [](const void* x, const void* y) {
-                //         return ( *(int*)x - *(int*)y );
+                //         return ( *(int64_t*)x - *(int64_t*)y );
                 //     });
                 
                 std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
@@ -91,12 +91,12 @@ TEST_CASE( "sort_pivots_1", "" )
     using namespace std;
 
     SECTION( "sort_pivots" ) {
-        vector< int > values = { 1, 2, 3, 4, 5, 7, 6 };
-        vector< int > pivots = {6};
+        vector< int64_t > values = { 1, 2, 3, 4, 5, 7, 6 };
+        vector< int64_t > pivots = {6};
         const auto results = sort_pivots( values, 0, values.size(), std::move(pivots) );
 
-        REQUIRE( values == vector<int>{ 6, 2, 3, 4, 5, 7, 1  } );
-        REQUIRE( results == vector<int>{ 0 } );
+        REQUIRE( values == vector<int64_t>{ 6, 2, 3, 4, 5, 7, 1  } );
+        REQUIRE( results == vector<int64_t>{ 0 } );
     }
 }
 
@@ -105,17 +105,17 @@ TEST_CASE( "sort_pivots_2", "" )
     using namespace std;
 
     SECTION( "all_permutations" ) {
-        vector< int > pivots = {5,6};
+        vector< int64_t > pivots = {5,6};
         do {
             //printList( pivots );
             
-            vector< int > values = { 1, 2, 3, 4, 5, 7, 6 };
+            vector< int64_t > values = { 1, 2, 3, 4, 5, 7, 6 };
             
             auto ps = pivots;
             const auto results = sort_pivots( values, 0, values.size(), std::move(ps) );
 
-            REQUIRE( values == vector<int>{ 6, 7, 3, 4, 5, 1, 2} );
-            REQUIRE( results == vector<int>{ 0, 1 } );
+            REQUIRE( values == vector<int64_t>{ 6, 7, 3, 4, 5, 1, 2} );
+            REQUIRE( results == vector<int64_t>{ 0, 1 } );
         } while( std::next_permutation(std::begin(pivots), std::end(pivots)) );
     }
 }
@@ -125,17 +125,17 @@ TEST_CASE( "sort_pivots_3", "" )
     using namespace std;
 
     SECTION( "all_permutations" ) {
-        vector< int > pivots = {0, 3, 6};
+        vector< int64_t > pivots = {0, 3, 6};
         do {
             //printList( pivots );
             
-            vector< int > values = { 4, 2, 3, 6, 5, 7, 1 };
+            vector< int64_t > values = { 4, 2, 3, 6, 5, 7, 1 };
 
             auto ps = pivots;
             const auto results = sort_pivots( values, 0, values.size(), std::move(ps) );
 
-            REQUIRE( values == vector<int>{ 1, 4, 6, 2, 5, 7, 3 } );
-            REQUIRE( results == vector<int>{ 0, 1, 2 } );
+            REQUIRE( values == vector<int64_t>{ 1, 4, 6, 2, 5, 7, 3 } );
+            REQUIRE( results == vector<int64_t>{ 0, 1, 2 } );
         } while( std::next_permutation(std::begin(pivots), std::end(pivots)) );
     }
 }
@@ -145,17 +145,17 @@ TEST_CASE( "sort_pivots_4", "" )
     using namespace std;
 
     SECTION( "all_permutations" ) {
-        vector< int > pivots = {0, 1, 2, 3};
+        vector< int64_t > pivots = {0, 1, 2, 3};
         do {
             //printList( pivots );
             
-            vector< int > values = { 4, 2, 3, 6, 5, 7, 1 };
+            vector< int64_t > values = { 4, 2, 3, 6, 5, 7, 1 };
 
             auto ps = pivots;
             const auto results = sort_pivots( values, 0, values.size(), std::move(ps) );
 
-            REQUIRE( values == vector<int>{ 2, 3, 4, 6, 5, 7, 1 } );
-            REQUIRE( results == vector<int>{ 0, 1, 2, 3 } );
+            REQUIRE( values == vector<int64_t>{ 2, 3, 4, 6, 5, 7, 1 } );
+            REQUIRE( results == vector<int64_t>{ 0, 1, 2, 3 } );
         } while( std::next_permutation(std::begin(pivots), std::end(pivots)) );
     }
 }
@@ -165,17 +165,17 @@ TEST_CASE( "sort_pivots_5", "" )
     using namespace std;
 
     SECTION( "all_permutations" ) {
-        vector< int > pivots = {2, 3, 4, 5, 6};
+        vector< int64_t > pivots = {2, 3, 4, 5, 6};
         do {
             //printList( pivots );
             
-            vector< int > values = { 4, 2, 3, 6, 5, 7, 1 };
+            vector< int64_t > values = { 4, 2, 3, 6, 5, 7, 1 };
 
             auto ps = pivots;
             const auto results = sort_pivots( values, 0, values.size(), std::move(ps) );
 
-            REQUIRE( values == vector<int>{ 1, 3, 5, 6, 7, 2, 4 } );
-            REQUIRE( results == vector<int>{ 0, 1, 2, 3, 4 } );
+            REQUIRE( values == vector<int64_t>{ 1, 3, 5, 6, 7, 2, 4 } );
+            REQUIRE( results == vector<int64_t>{ 0, 1, 2, 3, 4 } );
         } while( std::next_permutation(std::begin(pivots), std::end(pivots)) );
     }
 }
@@ -185,17 +185,17 @@ TEST_CASE( "sort_pivots_narrow_range_middle", "" )
     using namespace std;
 
     SECTION( "all_permutations" ) {
-        vector< int > pivots = {1, 2, 3};
+        vector< int64_t > pivots = {1, 2, 3};
         do {
             //printList( pivots );
             
-            vector< int > values = { 4, 6, 2, 3, 5, 7, 1 };
+            vector< int64_t > values = { 4, 6, 2, 3, 5, 7, 1 };
 
             auto ps = pivots;
             const auto results = sort_pivots( values, 0, 4, std::move(ps) );
 
-            REQUIRE( values == vector<int>{ 2, 3, 6, 4, 5, 7, 1 } );
-            REQUIRE( results == vector<int>{ 0, 1, 2 } );
+            REQUIRE( values == vector<int64_t>{ 2, 3, 6, 4, 5, 7, 1 } );
+            REQUIRE( results == vector<int64_t>{ 0, 1, 2 } );
         } while( std::next_permutation(std::begin(pivots), std::end(pivots)) );
     }
 }
@@ -205,17 +205,17 @@ TEST_CASE( "sort_pivots_narrow_range_begin", "" )
     using namespace std;
 
     SECTION( "all_permutations" ) {
-        vector< int > pivots = {1, 2, 3};
+        vector< int64_t > pivots = {1, 2, 3};
         do {
             //printList( pivots );
             
-            vector< int > values = { 4, 6, 2, 3, 5, 7, 1 };
+            vector< int64_t > values = { 4, 6, 2, 3, 5, 7, 1 };
 
             auto ps = pivots;
             const auto results = sort_pivots( values, 0, 3, std::move(ps) );
 
-            REQUIRE( values == vector<int>{ 2, 3, 6, 4, 5, 7, 1 } );
-            REQUIRE( results == vector<int>{ 0, 1, 2 } );
+            REQUIRE( values == vector<int64_t>{ 2, 3, 6, 4, 5, 7, 1 } );
+            REQUIRE( results == vector<int64_t>{ 0, 1, 2 } );
         } while( std::next_permutation(std::begin(pivots), std::end(pivots)) );
     }
 }
@@ -225,26 +225,26 @@ TEST_CASE( "sort_pivots_narrow_range_end", "" )
     using namespace std;
 
     SECTION( "all_permutations" ) {
-        vector< int > pivots = {4, 5, 6};
+        vector< int64_t > pivots = {4, 5, 6};
         do {
             //printList( pivots );
             
-            vector< int > values = { 4, 6, 2, 3, 5, 7, 1 };
+            vector< int64_t > values = { 4, 6, 2, 3, 5, 7, 1 };
 
             auto ps = pivots;
             const auto results = sort_pivots( values, 3, 6, std::move(ps) );
 
-            REQUIRE( values == vector<int>{ 4, 6, 2, 1, 5, 7, 3 } );
-            REQUIRE( results == vector<int>{ 3, 4, 5 } );
+            REQUIRE( values == vector<int64_t>{ 4, 6, 2, 1, 5, 7, 3 } );
+            REQUIRE( results == vector<int64_t>{ 3, 4, 5 } );
         } while( std::next_permutation(std::begin(pivots), std::end(pivots)) );
     }
 }
 
 // bool test_partitioning()
 // {
-//     for( int i = 0; i < 100; ++i )
+//     for( int64_t i = 0; i < 100; ++i )
 //     {
-//         std::vector< int > v = getRandomList( 50 );
+//         std::vector< int64_t > v = getRandomList( 50 );
 
 //         auto l = get_pivot( v, 0, v.size() - 1, 4 );
         
@@ -252,7 +252,7 @@ TEST_CASE( "sort_pivots_narrow_range_end", "" )
 
 //         for( auto pivot : l )
 //         {
-//             for( int j = 0; j <= pivot; ++j )
+//             for( int64_t j = 0; j <= pivot; ++j )
 //             {
 //                 if( v[pivot] < v[j] ) { printf( "Error" ); return false; }
 //             }
